refactored the folder structure for e2e tests
