--- e2e/tests/1v1Journey_test.js
@@ -1,77 +1,0 @@
-const config = require('../config.js');
-const {assignCaseRoleToUser, unAssignAllUsers, addUserCaseMapping} = require('../api/caseRoleAssignmentHelper');
-const {waitForFinishedBusinessProcess} = require('../api/testingSupport');
-
-const caseEventMessage = eventName => `Case ${caseNumber} has been updated with event: ${eventName}`;
-const caseId = () => `${caseNumber.split('-').join('').replace(/#/, '')}`;
-
-const claimant1 = {
-  litigantInPerson: true
-};
-
-const respondent1 = {
-  represented: true,
-  representativeRegistered: true,
-  representativeOrgNumber: 2
-};
-
-let caseNumber;
-
-Feature('1v1 - Claim Journey @e2e-unspec @e2e-multiparty @e2e-1v1');
-
-Scenario('Applicant solicitor creates claim @create-claim', async ({I}) => {
-  await I.login(config.applicantSolicitorUser);
-  await I.createCase(claimant1, null, respondent1, null);
-  caseNumber = await I.grabCaseNumber();
-  await I.see(`Case ${caseNumber} has been created.`);
-  await addUserCaseMapping(caseId(),config.applicantSolicitorUser);
-}).retry(3);
-
-Scenario('Applicant solicitor notifies defendant solicitor of claim', async ({I}) => {
-  await I.notifyClaim();
-  await I.see(caseEventMessage('Notify claim'));
-  await assignCaseRoleToUser(caseId(), 'RESPONDENTSOLICITORONE', config.defendantSolicitorUser);
-}).retry(3);
-
-Scenario('Applicant solicitor notifies defendant solicitor of claim details', async ({I}) => {
-  await I.notifyClaimDetails();
-  await I.see(caseEventMessage('Notify claim details'));
-  await I.click('Sign out');
-}).retry(3);
-
-Scenario('Defendant solicitor acknowledges claim', async ({I}) => {
-  await I.login(config.defendantSolicitorUser);
-  await I.acknowledgeClaim('fullDefence');
-  await I.see(caseEventMessage('Acknowledge claim'));
-  await I.click('Sign out');
-}).retry(3);
-
-Scenario('Defendant solicitor requests deadline extension', async ({I}) => {
-  await I.login(config.defendantSolicitorUser);
-  await I.informAgreedExtensionDate(1);
-  await I.see(caseEventMessage('Inform agreed extension date'));
-}).retry(3);
-
-Scenario('Defendant solicitor adds defendant litigation friend', async ({I}) => {
-  await I.login(config.defendantSolicitorUser);
-  await I.addDefendantLitigationFriend();
-  await I.see(caseEventMessage('Add litigation friend'));
-});
-
-Scenario('Defendant solicitor responds to claim', async ({I}) => {
-  await I.login(config.defendantSolicitorUser);
-  await I.respondToClaim({defendant1Response: 'fullDefence'});
-  await I.see(caseEventMessage('Respond to claim'));
-  await I.click('Sign out');
-}).retry(3);
-
-Scenario('Claimant solicitor responds to defence', async ({I}) => {
-  await I.login(config.applicantSolicitorUser);
-  await I.respondToDefence('ONE_V_ONE');
-  await I.see(caseEventMessage('View and respond to defence'));
-  await waitForFinishedBusinessProcess(caseId());
-}).retry(3);
-
-AfterSuite(async  () => {
-  await unAssignAllUsers();
-});
