+++ e2e/tests/api_tests/damages/api_1v1_test.js
@@ -1,9 +1,9 @@
 /* eslint-disable no-unused-vars */
 
-const config = require('../../config.js');
-const mpScenario = 'TWO_V_ONE';
+const config = require('../../../config.js');
+const mpScenario = 'ONE_V_ONE';
 
-Feature('CCD 1v2 Same Solicitor API test @api-unspec @api-multiparty @api-tests-2v1');
+Feature('CCD 1v1 API test @api-unspec @api-multiparty @api-tests-1v1');
 
 Scenario('Create claim', async ({I, api}) => {
   await api.createClaimWithRepresentedRespondent(config.applicantSolicitorUser, mpScenario);
@@ -38,11 +38,30 @@ Scenario('Inform agreed extension date', async ({I, api}) => {
 });
 
 Scenario('Defendant response', async ({I, api}) => {
-  await api.defendantResponse(config.defendantSolicitorUser, 'TWO_V_ONE');
+  await api.defendantResponse(config.defendantSolicitorUser, mpScenario);
 });
 
 Scenario('Claimant response', async ({I, api}) => {
   await api.claimantResponse(config.applicantSolicitorUser, mpScenario);
+});
+
+Scenario('Create claim where respondent is litigant in person', async ({I, api}) => {
+  await api.createClaimWithRespondentLitigantInPerson(config.applicantSolicitorUser);
+});
+
+Scenario('Create claim where respondent solicitor is not registered in my hmcts', async ({I, api}) => {
+  await api.createClaimWithRespondentSolicitorFirmNotInMyHmcts(config.applicantSolicitorUser);
+});
+
+Scenario('Create claim and move it to caseman', async ({I, api}) => {
+  await api.createClaimWithRepresentedRespondent(config.applicantSolicitorUser);
+  await api.moveCaseToCaseman(config.adminUser);
+});
+
+// This will be enabled when PAY-3817 issue of two minutes is fixed
+Scenario.skip('Resubmit claim after payment failure on PBA account ', async ({I, api}) => {
+  await api.createClaimWithFailingPBAAccount(config.applicantSolicitorUser);
+  await api.resubmitClaim(config.applicantSolicitorUser);
 });
 
 AfterSuite(async  ({api}) => {
