+++ e2e/tests/ui_tests/damages/1v1CreateClaim_test.js
@@ -1,16 +1,14 @@
-const config = require('../../config.js');
-const {waitForFinishedBusinessProcess} = require('../../api/testingSupport');
-const {addUserCaseMapping, assignCaseRoleToUser, unAssignAllUsers} = require('../../api/caseRoleAssignmentHelper');
+const config = require('../../../config.js');
+const {assignCaseRoleToUser, unAssignAllUsers, addUserCaseMapping} = require('../../../api/caseRoleAssignmentHelper');
+const {waitForFinishedBusinessProcess} = require('../../../api/testingSupport');
 
 const caseEventMessage = eventName => `Case ${caseNumber} has been updated with event: ${eventName}`;
 const caseId = () => `${caseNumber.split('-').join('').replace(/#/, '')}`;
 
 const claimant1 = {
-  litigantInPerson: false
+  litigantInPerson: true
 };
-const claimant2 = {
-  litigantInPerson: false
-};
+
 const respondent1 = {
   represented: true,
   representativeRegistered: true,
@@ -19,26 +17,23 @@ const respondent1 = {
 
 let caseNumber;
 
-Feature('2v1 Claim Journey @e2e-unspec @e2e-nightly @e2e-2v1');
+Feature('1v1 - Claim Journey @e2e-unspec @e2e-multiparty @e2e-1v1');
 
-Scenario('Claimant solicitor raises a claim for 2 claimants against 1 defendant', async ({I}) => {
+Scenario('Applicant solicitor creates claim @create-claim', async ({I}) => {
   await I.login(config.applicantSolicitorUser);
-  await I.createCase(claimant1, claimant2, respondent1, null);
+  await I.createCase(claimant1, null, respondent1, null);
   caseNumber = await I.grabCaseNumber();
   await I.see(`Case ${caseNumber} has been created.`);
-  addUserCaseMapping(caseId(), config.applicantSolicitorUser);
+  await addUserCaseMapping(caseId(),config.applicantSolicitorUser);
 }).retry(3);
 
-Scenario('Claimant solicitor notifies defendant of claim', async ({I}) => {
-  await I.login(config.applicantSolicitorUser);
+Scenario('Applicant solicitor notifies defendant solicitor of claim', async ({I}) => {
   await I.notifyClaim();
   await I.see(caseEventMessage('Notify claim'));
   await assignCaseRoleToUser(caseId(), 'RESPONDENTSOLICITORONE', config.defendantSolicitorUser);
-  await assignCaseRoleToUser(caseId(), 'APPLICANTSOLICITORTWO', config.secondDefendantSolicitorUser);
 }).retry(3);
 
-Scenario('Claimant solicitor notifies defendant solicitor of claim details', async ({I}) => {
-  await I.login(config.applicantSolicitorUser);
+Scenario('Applicant solicitor notifies defendant solicitor of claim details', async ({I}) => {
   await I.notifyClaimDetails();
   await I.see(caseEventMessage('Notify claim details'));
   await I.click('Sign out');
@@ -46,15 +41,14 @@ Scenario('Claimant solicitor notifies defendant solicitor of claim details', asy
 
 Scenario('Defendant solicitor acknowledges claim', async ({I}) => {
   await I.login(config.defendantSolicitorUser);
-  await I.acknowledgeClaim('fullDefence', null, 'fullDefence');
+  await I.acknowledgeClaim('fullDefence');
   await I.see(caseEventMessage('Acknowledge claim'));
   await I.click('Sign out');
 }).retry(3);
 
 Scenario('Defendant solicitor requests deadline extension', async ({I}) => {
   await I.login(config.defendantSolicitorUser);
-  await I.navigateToCaseDetails(caseId());
-  await I.informAgreedExtensionDate();
+  await I.informAgreedExtensionDate(1);
   await I.see(caseEventMessage('Inform agreed extension date'));
 }).retry(3);
 
@@ -62,23 +56,18 @@ Scenario('Defendant solicitor adds defendant litigation friend', async ({I}) => 
   await I.login(config.defendantSolicitorUser);
   await I.addDefendantLitigationFriend();
   await I.see(caseEventMessage('Add litigation friend'));
-}).retry(3);
+});
 
-Scenario('Defendants solicitor reject claim of both claimants', async ({I}) => {
+Scenario('Defendant solicitor responds to claim', async ({I}) => {
   await I.login(config.defendantSolicitorUser);
-  await I.navigateToCaseDetails('1642700629930139');
-  await I.respondToClaim({
-    twoDefendants: false,
-    defendant1Response: 'fullDefence',
-    defendant1ResponseToApplicant2: 'fullDefence'
-  });
+  await I.respondToClaim({defendant1Response: 'fullDefence'});
   await I.see(caseEventMessage('Respond to claim'));
   await I.click('Sign out');
 }).retry(3);
 
 Scenario('Claimant solicitor responds to defence', async ({I}) => {
   await I.login(config.applicantSolicitorUser);
-  await I.respondToDefence('TWO_V_ONE');
+  await I.respondToDefence('ONE_V_ONE');
   await I.see(caseEventMessage('View and respond to defence'));
   await waitForFinishedBusinessProcess(caseId());
 }).retry(3);
