+++ e2e/tests/ui_tests/damages/1v2CreateClaim_DiffSol_test.js
@@ -1,0 +1,103 @@
+const config = require('../../../config.js');
+const parties = require('../../../helpers/party');
+const {assignCaseRoleToUser, addUserCaseMapping, unAssignAllUsers} = require('../../../api/caseRoleAssignmentHelper');
+const {waitForFinishedBusinessProcess} = require('../../../api/testingSupport');
+
+const caseEventMessage = eventName => `Case ${caseNumber} has been updated with event: ${eventName}`;
+const caseId = () => `${caseNumber.split('-').join('').replace(/#/, '')}`;
+
+const claimant1 = {
+  litigantInPerson: false
+};
+const respondent1 = {
+  represented: true,
+  representativeRegistered: true,
+  representativeOrgNumber: 2
+};
+const respondent2 = {
+  represented: true,
+  sameLegalRepresentativeAsRespondent1: false,
+  representativeOrgNumber: 2
+};
+
+let caseNumber;
+
+Feature('1v2 Different Solicitors Claim Journey @e2e-unspec @e2e-nightly @e2e-1v2DS');
+
+Scenario('Claimant solicitor raises a claim against 2 defendants who have different solicitors', async ({I}) => {
+  await I.login(config.applicantSolicitorUser);
+  await I.createCase(claimant1, null, respondent1, respondent2);
+  caseNumber = await I.grabCaseNumber();
+  await I.see(`Case ${caseNumber} has been created.`);
+  addUserCaseMapping(caseId(), config.applicantSolicitorUser);
+}).retry(3);
+
+Scenario('Claimant solicitor notifies both defendants of claim', async ({I}) => {
+  await I.login(config.applicantSolicitorUser);
+  await I.notifyClaim('both');
+  await I.see(caseEventMessage('Notify claim'));
+  await assignCaseRoleToUser(caseId(), 'RESPONDENTSOLICITORONE', config.defendantSolicitorUser);
+  await assignCaseRoleToUser(caseId(),  'RESPONDENTSOLICITORTWO', config.secondDefendantSolicitorUser);
+}).retry(3);
+
+Scenario('Claimant solicitor notifies defendant solicitors of claim details', async ({I}) => {
+  await I.login(config.applicantSolicitorUser);
+  await I.notifyClaimDetails('both');
+  await I.see(caseEventMessage('Notify claim details'));
+  await I.click('Sign out');
+}).retry(3);
+
+Scenario('Defendant 1 solicitor acknowledges claim', async ({I}) => {
+  await I.login(config.defendantSolicitorUser);
+  await I.acknowledgeClaim('fullDefence');
+  await I.see(caseEventMessage('Acknowledge claim'));
+  await I.click('Sign out');
+}).retry(3);
+
+Scenario('Defendant 2 solicitor acknowledges claim', async ({I}) => {
+  await I.login(config.secondDefendantSolicitorUser);
+  await I.acknowledgeClaim(null, 'fullDefence');
+  await I.see(caseEventMessage('Acknowledge claim'));
+  await I.click('Sign out');
+}).retry(3);
+
+Scenario('Defendant 1 solicitor requests deadline extension', async ({I}) => {
+  await I.login(config.defendantSolicitorUser);
+  await I.navigateToCaseDetails(caseId());
+  await I.informAgreedExtensionDate();
+  await I.see(caseEventMessage('Inform agreed extension date'));
+}).retry(3);
+
+Scenario('Defendant 1 solicitor adds defendant litigation friend', async ({I}) => {
+  await I.login(config.defendantSolicitorUser);
+  await I.addDefendantLitigationFriend();
+  await I.see(caseEventMessage('Add litigation friend'));
+}).retry(3);
+
+Scenario('Defendant 1 solicitor rejects claim for defendant 1', async ({I}) => {
+  await I.login(config.defendantSolicitorUser);
+  await I.respondToClaim({
+    defendant1Response: 'fullDefence'});
+  await I.see(caseEventMessage('Respond to claim'));
+  await I.click('Sign out');
+}).retry(3);
+
+Scenario('Defendant 2 solicitor rejects claim for defendant 2', async ({I}) => {
+  await I.login(config.secondDefendantSolicitorUser);
+  await I.respondToClaim({
+    party: parties.RESPONDENT_SOLICITOR_2,
+    defendant2Response: 'fullDefence'});
+  await I.see(caseEventMessage('Respond to claim'));
+  await I.click('Sign out');
+}).retry(3);
+
+Scenario('Claimant solicitor responds to defence', async ({I}) => {
+  await I.login(config.applicantSolicitorUser);
+  await I.respondToDefence('ONE_V_TWO_TWO_LEGAL_REP');
+  await I.see(caseEventMessage('View and respond to defence'));
+  await waitForFinishedBusinessProcess(caseId());
+}).retry(3);
+
+AfterSuite(async  () => {
+  await unAssignAllUsers();
+});
