+++ e2e/tests/ui_tests/damages/1v2CreateClaim_offline_test.js
@@ -1,4 +1,4 @@
-const config = require('../../config.js');
+const config = require('../../../config.js');
 
 const claimant1 = {
   litigantInPerson: false
