+++ e2e/tests/ui_tests/damages/2v1CreateClaim_test.js
@@ -1,7 +1,6 @@
-const config = require('../../config.js');
-const parties = require('../../helpers/party');
-const {waitForFinishedBusinessProcess} = require('../../api/testingSupport');
-const {addUserCaseMapping, assignCaseRoleToUser, unAssignAllUsers} = require('../../api/caseRoleAssignmentHelper');
+const config = require('../../../config.js');
+const {waitForFinishedBusinessProcess} = require('../../../api/testingSupport');
+const {addUserCaseMapping, assignCaseRoleToUser, unAssignAllUsers} = require('../../../api/caseRoleAssignmentHelper');
 
 const caseEventMessage = eventName => `Case ${caseNumber} has been updated with event: ${eventName}`;
 const caseId = () => `${caseNumber.split('-').join('').replace(/#/, '')}`;
@@ -9,34 +8,33 @@ const caseId = () => `${caseNumber.split('-').join('').replace(/#/, '')}`;
 const claimant1 = {
   litigantInPerson: false
 };
+const claimant2 = {
+  litigantInPerson: false
+};
 const respondent1 = {
   represented: true,
   representativeRegistered: true,
   representativeOrgNumber: 2
 };
-const respondent2 = {
-  represented: true,
-  sameLegalRepresentativeAsRespondent1: true
-};
 
 let caseNumber;
 
-Feature('1v2 Same Solicitor Claim Journey @e2e-unspec @e2e-multiparty @e2e-1v2SS');
+Feature('2v1 Claim Journey @e2e-unspec @e2e-nightly @e2e-2v1');
 
-Scenario('Claimant solicitor raises a claim against 2 defendants who have the same solicitor', async ({I}) => {
+Scenario('Claimant solicitor raises a claim for 2 claimants against 1 defendant', async ({I}) => {
   await I.login(config.applicantSolicitorUser);
-  await I.createCase(claimant1, null, respondent1, respondent2);
+  await I.createCase(claimant1, claimant2, respondent1, null);
   caseNumber = await I.grabCaseNumber();
   await I.see(`Case ${caseNumber} has been created.`);
   addUserCaseMapping(caseId(), config.applicantSolicitorUser);
 }).retry(3);
 
-Scenario('Claimant solicitor notifies both defendants of claim', async ({I}) => {
+Scenario('Claimant solicitor notifies defendant of claim', async ({I}) => {
   await I.login(config.applicantSolicitorUser);
   await I.notifyClaim();
   await I.see(caseEventMessage('Notify claim'));
   await assignCaseRoleToUser(caseId(), 'RESPONDENTSOLICITORONE', config.defendantSolicitorUser);
-  await assignCaseRoleToUser(caseId(), 'RESPONDENTSOLICITORTWO', config.secondDefendantSolicitorUser);
+  await assignCaseRoleToUser(caseId(), 'APPLICANTSOLICITORTWO', config.secondDefendantSolicitorUser);
 }).retry(3);
 
 Scenario('Claimant solicitor notifies defendant solicitor of claim details', async ({I}) => {
@@ -48,7 +46,7 @@ Scenario('Claimant solicitor notifies defendant solicitor of claim details', asy
 
 Scenario('Defendant solicitor acknowledges claim', async ({I}) => {
   await I.login(config.defendantSolicitorUser);
-  await I.acknowledgeClaim('fullDefence', 'fullDefence', null, true);
+  await I.acknowledgeClaim('fullDefence', null, 'fullDefence');
   await I.see(caseEventMessage('Acknowledge claim'));
   await I.click('Sign out');
 }).retry(3);
@@ -62,24 +60,25 @@ Scenario('Defendant solicitor requests deadline extension', async ({I}) => {
 
 Scenario('Defendant solicitor adds defendant litigation friend', async ({I}) => {
   await I.login(config.defendantSolicitorUser);
-  await I.addDefendantLitigationFriend('both');
+  await I.addDefendantLitigationFriend();
   await I.see(caseEventMessage('Add litigation friend'));
 }).retry(3);
 
-Scenario('Defendants solicitor rejects claim for both defendants', async ({I}) => {
+Scenario('Defendants solicitor reject claim of both claimants', async ({I}) => {
   await I.login(config.defendantSolicitorUser);
+  await I.navigateToCaseDetails('1642700629930139');
   await I.respondToClaim({
-    party: parties.RESPONDENT_SOLICITOR_1,
-    twoDefendants: true,
-    sameResponse: true,
-    defendant1Response: 'fullDefence'});
+    twoDefendants: false,
+    defendant1Response: 'fullDefence',
+    defendant1ResponseToApplicant2: 'fullDefence'
+  });
   await I.see(caseEventMessage('Respond to claim'));
   await I.click('Sign out');
 }).retry(3);
 
 Scenario('Claimant solicitor responds to defence', async ({I}) => {
   await I.login(config.applicantSolicitorUser);
-  await I.respondToDefence('ONE_V_TWO_ONE_LEGAL_REP');
+  await I.respondToDefence('TWO_V_ONE');
   await I.see(caseEventMessage('View and respond to defence'));
   await waitForFinishedBusinessProcess(caseId());
 }).retry(3);
