+++ e2e/tests/ui_tests/sdo/sdo_journey_test.js
@@ -1,4 +1,4 @@
-const config = require('../../config.js');
+const config = require('../../../config.js');
 
 const claimant1 = {
   litigantInPerson: true
