+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "lint": "eslint e2e",
     "test": "npx codeceptjs run --steps",
-    "test:smoke": "MOCHAWESOME_REPORTFILENAME=smoke npx codeceptjs run --grep @smoke-tests --reporter mocha-multi --verbose",
+    "test:smoke": "MOCHAWESOME_REPORTFILENAME=smoke npx codeceptjs run --grep @smoke-tests-unspec --reporter mocha-multi --verbose",
     "test:smoke-spec": "MOCHAWESOME_REPORTFILENAME=smoke npx codeceptjs run --grep @smoke-tests-spec --reporter mocha-multi --verbose",
     "test:api-unspec": "MOCHAWESOME_REPORTFILENAME=api npx codeceptjs run --grep @api-unspec --reporter mocha-multi --verbose",
     "test:e2e-unspec": "MOCHAWESOME_REPORTFILENAME=api npx codeceptjs run --grep @e2e-unspec --reporter mocha-multi --verbose",
